changed from uint to size_t for opencv Vector
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@19305 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: c53f3c6e0bca12b831e7a2a70e61ca060072d981

+++ sandbox/descriptors_3d/src/bounding_box.cpp
@@ -242,7 +242,7 @@ void BoundingBox::computeBoundingBoxFeatures(const robot_msgs::PointCloud& data,
   if (nbr_pts == 0)
   {
     ROS_WARN("BoundingBox::computeBoundingBoxFeatures() Passed empty bounding box indices");
-    for (unsigned int i = 0 ; i < result_size_ ; i++)
+    for (size_t i = 0 ; i < result_size_ ; i++)
     {
       result[i] = 0.0;
     }
